agentes-ia/tenants-admin/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-secret-manager
